+++ tee-worker/omni-executor/executor-storage/src/lib.rs
@@ -25,8 +25,8 @@ pub use pumpx_jwt::PumpxJwtStorage;
 mod intent_id;
 pub use intent_id::IntentIdStorage;
 mod asset_lock;
-mod omni_account_profile;
-pub use omni_account_profile::OmniAccountProfileStorage;
+mod pumpx_account_profile;
+pub use pumpx_account_profile::PumpxProfileStorage;
 
 pub use asset_lock::AssetLockStorage;
 pub use asset_lock::Key as AssetLockStorageKey;
